fix(mc-host): address fourth-round review findings
- forced shutdown no longer races bind completion: bind callbacks and
  their open_route wrapper are abort-exempt and self-bounded, the
  forced sweep defers mid-bind routes to that wrapper (marking them
  close-requested) instead of finalizing slots a late install_bound or
  take_rejected_bind would panic against, and the post-force wait
  covers the bind-then-route-gone chain before the lock releases
- an oversized stream item settles the request as internal_error so a
  truncated stream can never end in a successful StreamEnd, and a
  unary Response after emitted StreamData normalizes to internal_error
- a dead generation's routes close concurrently (bounded by
  max_routes) instead of multiplying route-gone latency serially
- startup validates the manifest module_id against route.open's field
  constraints so the catalog never advertises an unbindable module
- cast stat identity fields for platforms where st_dev is not u64
- test fixes: the settlement-race permit is consumed with forget() so
  every iteration races its release, and the wake.create catalog probe
  compares windows of the needle's actual length

+++ crates/mc-host/src/instance.rs
@@ -288,7 +288,10 @@ impl InstanceGuard {
         if !is_secure_regular(&stat) {
             return;
         }
-        if stat.st_dev != identity.dev || stat.st_ino != identity.ino {
+        // `Stat` field types vary by platform (macOS `st_dev` is `i32`); the
+        // casts are no-ops on Linux but load-bearing elsewhere.
+        #[allow(clippy::unnecessary_cast)]
+        if stat.st_dev as u64 != identity.dev || stat.st_ino as u64 != identity.ino {
             return;
         }
         let Ok(bytes) = read_all_fd(&fd, 65_536) else {
